+++ src/common/index.ts
@@ -2,8 +2,9 @@ import { RectGrid } from "./rectGrid";
 import { reviver, replacer } from "./serialization";
 import { shuffle } from "./shuffle";
 import { UserFacingError } from "./errors";
+import { HexTriGraph, SnubSquareGraph, SquareOrthGraph, SquareGraph } from "./graphs";
 
-export { RectGrid, reviver, replacer, shuffle, UserFacingError };
+export { RectGrid, reviver, replacer, shuffle, UserFacingError, HexTriGraph, SnubSquareGraph, SquareOrthGraph, SquareGraph };
 
 export type DirectionsCardinal = "N" | "E" | "S" | "W";
 export type DirectionsDiagonal = "NE" | "SE" | "SW" | "NW";