+++ src/games/index.ts
@@ -14,6 +14,7 @@ import { CephalopodGame, ICephalopodState } from "./ceph";
 import { LinesOfActionGame, ILinesOfActionState } from "./loa";
 import { PikemenGame, IPikemenState } from "./pikemen";
 import { OrdoGame, IOrdoState } from "./ordo";
+import { AttangleGame, IAttangleState } from "./attangle";
 
 export {
     APGamesInformation, GameBase, IAPGameState,
@@ -31,16 +32,17 @@ export {
     LinesOfActionGame, ILinesOfActionState,
     PikemenGame, IPikemenState,
     OrdoGame, IOrdoState,
+    AttangleGame, IAttangleState,
 };
 
 const games = new Map<string, typeof AmazonsGame | typeof BlamGame | typeof CannonGame |
                               typeof MchessGame | typeof HomeworldsGame | typeof EntropyGame |
                               typeof VolcanoGame | typeof MvolcanoGame | typeof ChaseGame |
                               typeof AbandeGame | typeof CephalopodGame | typeof LinesOfActionGame |
-                              typeof PikemenGame | typeof OrdoGame>();
+                              typeof PikemenGame | typeof OrdoGame | typeof AttangleGame>();
 // Manually add each game to the following array
 [AmazonsGame, BlamGame, CannonGame, MchessGame, HomeworldsGame, EntropyGame, VolcanoGame, MvolcanoGame,
-    ChaseGame, AbandeGame, CephalopodGame, LinesOfActionGame, PikemenGame, OrdoGame].forEach((g) => {
+    ChaseGame, AbandeGame, CephalopodGame, LinesOfActionGame, PikemenGame, OrdoGame, AttangleGame].forEach((g) => {
     if (games.has(g.gameinfo.uid)) {
         throw new Error("Another game with the UID '" + g.gameinfo.uid + "' has already been used. Duplicates are not allowed.");
     }
@@ -78,6 +80,8 @@ export function GameFactory(game: string, ...args: any[]): GameBase|undefined {
             return new PikemenGame(...args);
         case "ordo":
             return new OrdoGame(...args);
+        case "attangle":
+            return new AttangleGame(...args);
     }
     return;
 }