+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added Lines of Action. It's the 9x9 black hole variant, with an optional Scrambled Eggs initial layout. Supports move generation and stupid AI.
 - Added Pikemen, with move generation and brain-dead AI.
 - Added Ordo, with move generation and very, very slow AI (need to optimize move generation).
+- Added Attangle, with move generation and AI.
 
 ## [0.3.0] - 2021-11-12